change structrue folder + add eslint + add sequelize  migrate
