+++ config/database.js
@@ -1,8 +1,9 @@
-const { Sequelize } = require('sequelize');
-require('dotenv').config();
+const { Sequelize } = require("sequelize");
+require("dotenv").config();
 
-module.exports = new Sequelize('forum', 'root', process.env.DB_PW, {
-  host: 'localhost',
-  dialect: 'mysql'
+
+module.exports = new Sequelize("forum", "root", process.env.DB_PW, {
+	host: "localhost",
+	dialect: "mysql"
 });
